{"schema":"decodex/commit/1","summary":"Refine state runtime record module boundaries","authority":"manual"}
